realm: more work on support for background work profiling
