+++ src/realm/cuda/cuda_internal.h
@@ -43,7 +43,6 @@
 #include "realm/proc_impl.h"
 #include "realm/mem_impl.h"
 #include "realm/bgwork.h"
-#include "realm/bgwork_profile.h"
 #include "realm/transfer/channel.h"
 #include "realm/transfer/ib_memory.h"
 #include "realm/cuda/cuda_memcpy.h"
@@ -325,7 +324,7 @@ namespace Realm {
       void start_background_thread(Realm::CoreReservationSet &crs, size_t stack_size);
       void shutdown_background_thread(void);
 
-      bool do_work(TimeLimit work_until);
+      bool do_work(TimeLimit work_until, BgWorkProfileState &profstate);
 
     public:
       void thread_main(void);
