+++ src/realm/deppart/partitions.h
@@ -207,7 +207,7 @@ namespace Realm {
     void worker_thread_loop(void);
 
     // called by BackgroundWorkers
-    bool do_work(TimeLimit work_until);
+    bool do_work(TimeLimit work_until, BgWorkProfileState &profstate);
 
   protected:
     atomic<bool> shutdown_flag;
